Symbolize option keys sent to ::Redis.new
I think this will save some people the trouble of forgetting to
symbolize the keys to redis in the config.yml.  Copied the method
from activesupport rather than adding the dependency.

Also, tried to unit test it but it was too sticky and I didn't
want to be presumptuous about pulling in a mock redis library.
That's probably worth its own issue.